+++ R/mlr3verse_info.R
@@ -7,6 +7,9 @@
 #' @examples
 #' mlr3verse_info()
 mlr3verse_info = function() {
-  pkgs = c("mlr3", "mlr3tuning", "mlr3pipelines", "paradox", "mlr3filters", "mlr3db", "mlr3learners")
-  data.frame(Package = pkgs, Version = mlr3misc::map_chr(pkgs, function(x) as.character(packageVersion(x))))
+  pkgs = c("mlr3", "mlr3tuning", "mlr3pipelines", "paradox", "mlr3filters",
+    "mlr3db", "mlr3learners", "mlr3viz")
+  data.frame(Package = pkgs, Version = mlr3misc::map_chr(pkgs, function(x) {
+    as.character(packageVersion(x))
+  }))
 }